+++ src/app/controllers/ContactController.js
@@ -148,6 +148,12 @@ class ContactController {
   async delete(req, res) {
     const { id } = req.params;
 
+    if (!isValidUUID(id)) {
+      return res
+        .status(400)
+        .json({ error: true, message: STATUS.ALL.INVALID_USER_ID });
+    }
+
     await ContactsRepository.delete(id);
 
     res.status(200).json({ status: "Ok", message: STATUS.ALL.DELETED_ITEM });
